+++ src/botlib/be_ai_char.cpp
@@ -350,7 +350,7 @@ int BotFindCachedCharacter(char *charfile, float skill)
 	{
 		if ( !botcharacters[handle] ) continue;
 		if ( strcmp( botcharacters[handle]->filename, charfile ) == 0 &&
-			(skill < 0 || fabs(botcharacters[handle]->skill - skill) < 0.01) )
+			(skill < 0 || fabsf(botcharacters[handle]->skill - skill) < 0.01f) )
 		{
 			return handle;
 		} //end if
