+++ src/game/bg_misc.c
@@ -1264,7 +1264,7 @@ float vectoyaw( const vec3_t vec ) {
 		yaw = 0;
 	} else {
 		if (vec[PITCH]) {
-			yaw = ( atan2( vec[YAW], vec[PITCH]) * 180 / (float)M_PI );
+			yaw = ( atan2f( vec[YAW], vec[PITCH]) * 180 / (float)M_PI );
 		} else if (vec[YAW] > 0) {
 			yaw = 90;
 		} else {
@@ -1766,7 +1766,7 @@ void BG_EvaluateTrajectory( const trajectory_t *tr, int atTime, vec3_t result ) 
 		break;
 	case TR_SINE:
 		deltaTime = ( atTime - tr->trTime ) / (float) tr->trDuration;
-		phase = sin( deltaTime * M_PI * 2 );
+		phase = sinf( deltaTime * M_PI * 2 );
 		VectorMA( tr->trBase, phase, tr->trDelta, result );
 		break;
 	case TR_LINEAR_STOP:
@@ -1815,7 +1815,7 @@ void BG_EvaluateTrajectoryDelta( const trajectory_t *tr, int atTime, vec3_t resu
 		break;
 	case TR_SINE:
 		deltaTime = ( atTime - tr->trTime ) / (float) tr->trDuration;
-		phase = cos( deltaTime * M_PI * 2 );	// derivative of sin = cos
+		phase = cosf( deltaTime * M_PI * 2 );	// derivative of sin = cos
 		phase *= 0.5f;
 		VectorScale( tr->trDelta, phase, result );
 		break;
