+++ src/qcommon/cm_polylib.cpp
@@ -186,7 +186,7 @@ winding_t *BaseWindingForPlane (vec3_t normal, vec_t dist)
 	x = -1;
 	for (i=0 ; i<3; i++)
 	{
-		v = fabs(normal[i]);
+		v = fabsf(normal[i]);
 		if (v > max)
 		{
 			x = i;
